+++ CalendarView/CalendarViewList.php
@@ -11,27 +11,36 @@
  */
 
 namespace Zikula\PostCalendarModule\CalendarView;
-class CalendarViewWeek extends AbstractDays
+
+use \System;
+use CategoryRegistryUtil;
+use \SecurityUtil;
+use \UserUtil;
+use \LogUtil;
+use \ModUtil;
+use \CategoryUtil;
+use \DateTime;
+use \DataUtil;
+use \ZLanguage;
+use DateInterval;
+use DatePeriod;
+
+class CalendarViewList extends AbstractDays
 {
 
     /**
-     * An array of values to use in the computation of DateTime objects 
-     * Do not translate!
+     * How many months to view
      * 
-     * @var array 
+     * @var integer
      */
-    private $dayMap = array(
-        self::SUNDAY_IS_FIRST => "Sunday",
-        self::MONDAY_IS_FIRST => "Monday",
-        self::SATURDAY_IS_FIRST => "Saturday",
-    );
+    protected $listMonths = 1;
 
     /**
      * Set the cacheTag 
      */
     protected function setCacheTag()
     {
-        $this->cacheTag = $this->requestedDate->format('Ym');
+        $this->cacheTag = $this->requestedDate->format('Ymd');
     }
 
     /**
@@ -39,7 +48,7 @@ class CalendarViewWeek extends AbstractDays
      */
     protected function setTemplate()
     {
-        $this->template = 'user/week.tpl';
+        $this->template = 'user/list.tpl';
     }
 
     /**
@@ -48,14 +57,22 @@ class CalendarViewWeek extends AbstractDays
     protected function setDates()
     {
         $this->startDate = clone $this->requestedDate;
-        if ($this->requestedDate->format('w') <> $this->firstDayOfWeek) {
-            $this->startDate
-                    ->modify("last " . $this->dayMap[$this->firstDayOfWeek]);
-        }
         $this->endDate = clone $this->requestedDate;
         $this->endDate
-                ->modify("next " . $this->dayMap[$this->firstDayOfWeek])
-                ->modify("-1 day");
+                ->modify("+" . $this->listMonths . " months");
+
+        $interval = new DateInterval("P1D");
+        $datePeriod = new DatePeriod($this->startDate, $interval, $this->endDate);
+        $i = 0;
+        $week = 0;
+        foreach ($datePeriod as $date) {
+            $this->dateGraph[$week][$i] = $date->format('Y-m-d');
+            $i++;
+            if ($i > 6) {
+                $i = 0;
+                $week++;
+            }
+        }
     }
 
     /**
@@ -63,20 +80,19 @@ class CalendarViewWeek extends AbstractDays
      */
     protected function setup()
     {
-        $this->viewtype = 'week';
+        $this->viewtype = 'list';
+        $this->listMonths = ModUtil::getVar('ZikulaPostCalendarModule', 'pcListMonths');
 
         $prevClone = clone $this->requestedDate;
-        $prevClone->modify("last " . $this->dayMap[$this->firstDayOfWeek]);
-        if ($this->requestedDate->format('w') <> $this->firstDayOfWeek) {
-            $prevClone->modify("-7 days");
-        }
+        $prevClone->modify("-" . $this->listMonths . " months");
         $this->navigation['previous'] = ModUtil::url('PostCalendar', 'user', 'display', array(
                     'viewtype' => $this->viewtype,
                     'date' => $prevClone->format('Ymd'),
                     'userfilter' => $this->userFilter,
                     'filtercats' => $this->selectedCategories));
         $nextClone = clone $this->requestedDate;
-        $nextClone->modify("next " . $this->dayMap[$this->firstDayOfWeek]);
+        $nextClone->modify("+" . $this->listMonths . " months")
+                ->modify("+1 day");
         $this->navigation['next'] = ModUtil::url('PostCalendar', 'user', 'display', array(
                     'viewtype' => $this->viewtype,
                     'date' => $nextClone->format('Ymd'),
@@ -92,9 +108,9 @@ class CalendarViewWeek extends AbstractDays
     {
         if (!$this->isCached()) {
             // Load the events
-            $eventsByDate = ModUtil::apiFunc('PostCalendar', 'event', 'getEvents', array(
-                        'start' => clone $this->startDate,
-                        'end' => clone $this->endDate,
+            $eventsByDate = ModUtil::apiFunc('ZikulaPostCalendarModule', 'event', 'getEvents', array(
+                        'start' => $this->startDate,
+                        'end' => $this->endDate,
                         'filtercats' => $this->selectedCategories,
                         'date' => $this->requestedDate,
                         'userfilter' => $this->userFilter));
@@ -102,10 +118,11 @@ class CalendarViewWeek extends AbstractDays
             $this->view
                     ->assign('navBar', $this->navBar)
                     ->assign('navigation', $this->navigation)
+                    ->assign('dayDisplay', $this->dayDisplay)
+                    ->assign('graph', $this->dateGraph)
                     ->assign('eventsByDate', $eventsByDate)
-                    ->assign('startDate', $this->startDate->format('Ymd'))
-                    ->assign('endDate', $this->endDate->format('Ymd'))
-                    ->assign('requestedDate', $this->requestedDate->format('Ymd'));
+                    ->assign('startDate', $this->startDate)
+                    ->assign('endDate', $this->endDate);
         }
         return $this->view->fetch($this->template);
     }
