+++ Resources/public/js/postcalendar-function-updatefields.js
